requirements.txt
django==3.2.9
djangorestframework==3.12.4
python-decouple==3.5
django-filter-21.1
